Modeling linker steps as functions instead of classes. Changing visit() method to receive a visitor object instead of separated functions. Splitted linker into files. Moving particular Wollok scoping/linking definitions into a specific file to reduce complexity from the linker itself
