+++ src/model/visiting.js
@@ -3,20 +3,20 @@ import winston from 'winston'
 // this should be in the linker !
 const ignoredKeys = ['parent', 'link']
 
-export const visit = (node, fn, after = () => {}, parent) => {
+export const visit = (node, { onNode, afterNode = () => {} }, parent) => {
   if (!node.nodeType) { return node }
   winston.silly(`visiting ${node.nodeType}`)
-  fn(node, parent)
+  onNode(node, parent)
   Object.keys(node).forEach(key => {
     if (ignoredKeys.includes(key)) return
     const value = node[key]
     const list = Array.isArray(value) ? value : [value]
     list.filter(e => e.nodeType).forEach((e, i) => {
       winston.silly(`\tvisiting ${node.nodeType}.${key}[${i}]`)
-      visit(e, fn, after, node)
+      visit(e, { onNode, afterNode }, node)
     })
   })
-  after(node)
+  afterNode(node)
   return node
 }
 
@@ -35,10 +35,12 @@ function methodByConvention(object, node, preffix) {
 export const queryNodeByType = (root, type, filter = () => true) => {
   const possibles = []
   const matches = []
-  visit(root, node => {
-    possibles.push(node.nodeType)
-    if (node.nodeType === type && filter(node)) {
-      matches.push(node)
+  visit(root, {
+    onNode(node) {
+      possibles.push(node.nodeType)
+      if (node.nodeType === type && filter(node)) {
+        matches.push(node)
+      }
     }
   })
   if (matches.length === 0) {
